EventPro-backend/python_backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
pydantic
xendit
requests
firebase-admin